refactor: made major refactor
