+++ src/app/page.tsx
@@ -8,11 +8,10 @@ import Header from "@/components/layouts/Header";
 import Page from "@/components/layouts/Page";
 import MainMenu from "@/features/common/menu";
 import SettingModal from "@/features/common/SettingModal";
-import ArchivedTodoList from "@/features/routes/todo/ArchivedTodoList";
-import EditTodoModal from "@/features/routes/todo/EditTodoModal";
-import NewTodoModal from "@/features/routes/todo/NewTodoModal";
-import TodoDetailModal from "@/features/routes/todo/todoDetailModal";
-import TodoList from "@/features/routes/todo/TodoList";
+import TodoList from "@/features/routes/todoList";
+import EditTodoModal from "@/features/routes/todoModal/editTodoModal";
+import NewTodoModal from "@/features/routes/todoModal/newTodoModal";
+import TodoDetailModal from "@/features/routes/todoModal/todoDetailModal";
 import { useControlMenu } from "@/hooks/useMenu";
 import { useSetModal } from "@/hooks/useModals";
 import { useGetIsArchived } from "@/hooks/useTodo";
@@ -52,7 +51,7 @@ export default function Home() {
                 </div>
                 <Content>
                     <div style={{ height: `calc(100vh - ${headerHeight}px - ${footerHeight}px)` }}>
-                        {isArchived() ? <ArchivedTodoList /> : <TodoList />}
+                        <TodoList />
                     </div>
                 </Content>
                 <div ref={footerRef}>
