+++ src/features/routes/todoModal/newTodoModal/index.tsx
@@ -1,0 +1,69 @@
+import { useState } from "react";
+
+import Modal from "@/components/elements/Modal";
+import Content from "@/components/layouts/Content";
+import Footer from "@/components/layouts/Footer";
+import Header from "@/components/layouts/Header";
+import Page from "@/components/layouts/Page";
+import { useGetModal, useSetModal } from "@/hooks/useModals";
+import { CheckList, Todo, useSetNewTodo } from "@/hooks/useTodo";
+
+import CheckBox from "../checkBox";
+import TextAreaInput from "../TextAreaInput";
+import TextInput from "../TextInput";
+import FooterContent from "./FooterContent";
+import HeaderContent from "./HeaderContent";
+
+export default function NewTodoModal() {
+    const modal = useGetModal();
+    const setModal = useSetModal();
+    const addTodo = useSetNewTodo();
+
+    const [title, setTitle] = useState("");
+    const [content, setContent] = useState("");
+    const [useCheckBox, setUseCheckBox] = useState<boolean>(false);
+    const [checkList, setCheckList] = useState<CheckList[]>([]);
+    const [isPinned, setIsPinned] = useState<boolean>(false);
+
+    const confirmCreate = () => {
+        const id = "id" + Math.random().toString(16).slice(2);
+        const createdDate = new Date();
+        const newTodo: Todo = {
+            id: id,
+            title: title,
+            content: content,
+            checkList: checkList,
+            createdAt: createdDate,
+            lastEditAt: null,
+            isChecked: false,
+            isPinned: isPinned,
+            isArchived: false,
+        };
+        addTodo(newTodo);
+        setModal(null);
+        setTitle("");
+        setContent("");
+    };
+
+    return (
+        <Modal isOpen={modal === "newTodo"}>
+            <Page>
+                <Header>
+                    <HeaderContent isPinned={isPinned} setIsPinned={setIsPinned} />
+                </Header>
+                <Content>
+                    <main className="relative flex grow flex-col overflow-y-auto p-3">
+                        <TextInput label="Title" subText="Within 40 letters" maxLength={40} value={title} setValue={setTitle} />
+                        <TextAreaInput label="content" subText="" maxLength={undefined} value={content} setValue={setContent} />
+                        <div className="relative flex w-full flex-col">
+                            {useCheckBox && <CheckBox checkList={checkList} setCheckList={setCheckList} />}
+                        </div>
+                    </main>
+                </Content>
+                <Footer>
+                    <FooterContent useCheckBox={useCheckBox} setUseCheckBox={setUseCheckBox} confirmCreate={confirmCreate} />
+                </Footer>
+            </Page>
+        </Modal>
+    );
+}
