--- src/hooks/useList.ts
@@ -1,3 +1,0 @@
-export function useAddNewList() {
-    
-}
